+++ src/Files.Uwp/Views/CustomFolderIcons.xaml
@@ -3,7 +3,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:UserControls="using:Files.UserControls"
-    xmlns:common="using:Files.Common"
+    xmlns:shared="using:Files.Shared"
     xmlns:converters="using:Microsoft.Toolkit.Uwp.UI.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:helpers="using:Files.Helpers"
@@ -95,7 +95,7 @@
                     </ItemsPanelTemplate>
                 </GridView.ItemsPanel>
                 <GridView.ItemTemplate>
-                    <DataTemplate x:DataType="common:IconFileInfo">
+                    <DataTemplate x:DataType="shared:IconFileInfo">
                         <Grid Width="32" Height="32">
                             <Image UserControls:ImageFromBytes.SourceBytes="{x:Bind IconDataBytes}" />
                         </Grid>
